+++ src/http-message/src/ContentType.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Swoft\Http\Message;
+
+/**
+ * Class ContentType
+ * @package Swoft\Http\Message
+ */
+final class ContentType
+{
+    public const HTML = 'text/html';
+    public const TEXT = 'text/plain';
+    public const JSON = 'application/json';
+    public const XML  = 'application/xml';
+}