+++ src/http-message/src/Contract/ResponseInterface.php
@@ -1,0 +1,12 @@
+<?php declare(strict_types=1);
+
+namespace Swoft\Http\Message\Contract;
+
+/**
+ * Interface ResponseInterface
+ * @since 2.0
+ */
+interface ResponseInterface extends \Psr\Http\Message\ResponseInterface
+{
+
+}