+++ src/http-server/src/AutoLoader.php
@@ -2,6 +2,7 @@
 
 namespace Swoft\Http\Server;
 
+use Swoft\Http\Message\ContentType;
 use Swoft\Http\Server\Formatter\HtmlResponseFormatter;
 use Swoft\Http\Server\Formatter\JsonResponseFormatter;
 use Swoft\Http\Server\Formatter\XmlResponseFormatter;
@@ -40,8 +41,8 @@ class AutoLoader extends \Swoft\AutoLoader
         return [
             'httpRequest'     => [
                 'parsers' => [
-                    Request::CONTENT_JSON => bean(JsonRequestParser::class),
-                    Request::CONTENT_XML  => bean(XmlRequestParser::class),
+                    ContentType::XML  => bean(XmlRequestParser::class),
+                    ContentType::JSON => bean(JsonRequestParser::class),
                 ]
             ],
             'httpResponse'    => [