+++ src/http-server/src/Parser/JsonRequestParser.php
@@ -1,10 +1,9 @@
 <?php declare(strict_types=1);
 
-
 namespace Swoft\Http\Server\Parser;
 
-
 use Swoft\Bean\Annotation\Mapping\Bean;
+use Swoft\Http\Message\Contract\RequestParserInterface;
 use Swoft\Stdlib\Helper\JsonHelper;
 
 /**